refactor(indexer/handlers): extract fpmm/factory.ts (PoolDeployed + Transfer + applyLiquidityPositionDelta) (PR-S1)
First slice of the handlers/fpmm.ts split (1028 lines → 773). Move
into the new indexer-envio/src/handlers/fpmm/factory.ts (270 lines):

- applyLiquidityPositionDelta helper (called by Transfer here; future
  S2 liquidity.ts imports it for Mint + Burn)
- FPMMFactory.PoolDeployed handler — pool creation + dynamic fee-token
  registration via the SECURITY GATE (only ERC20s present in
  isKnownFeeToken get the Transfer handler installed)
- FPMM.Transfer handler — LP token ownership transfers

EventHandlers.ts is updated to import the new module so Envio sees the
side-effect handler registrations at module load.

Pure extraction. 380 tests pass.
